Create README.md
TrendingMovies - My first React project Hi, welcome and thanks for checking out my github page. 1 - This was part of a uni project. The task was to design & develop a full stack multi-page web application. the application should be functionally working and smooth to interact with for the user across modern browsers. It should be well designed and feel professional.

2 - This React project fetches data from an API, writes the data on a firebase database and then displays that data.

3 - I have disabled the database WRITE, and removed my API keys. you can use your own firebase database at (index.js) and your your own TMBD API keys at (components/FetchData.js). #Mohamed Ali